Update tests to use chai & be more readable
